--- a/build/Dockerfile
+++ b/build/Dockerfile
@@ -4,7 +4,7 @@ WORKDIR /usr/local/src
 
 RUN BUILD_DEPS="make gcc"
 RUN apt-get update \
-    && apt-get install -y --no-install-recommends "${BUILD_DEPS}" libmagickwand-dev ncurses-dev
+    && apt-get install -y --no-install-recommends "${BUILD_DEPS}" libmagickwand-dev ncurses-dev jq
 
 RUN git clone --depth 1 https://github.com/denilsonsa/img2xterm \
     && (cd img2xterm && make && make install) \
@@ -28,7 +28,7 @@ RUN go run /usr/local/src/src/bin/convert/png_convert.go \
         -to /tmp/cows/ \
         -padding 4 \
         -skip '["resources/", "misc/", "icons/", "items/", "items-outline/"]' \
-    && cp /tmp/original/pokesprite/data/pokemon.json /tmp/cows/ \
+    && cat /tmp/original/pokesprite/data/pokemon.json | jq -c .[] > /tmp/cows/pokemon.json \
     && rm -rf /tmp/original/pokesprite
 
 ADD . .