--- a/build/scripts/build_assets.sh
+++ b/build/scripts/build_assets.sh
@@ -6,6 +6,7 @@ FROM="${1:-/tmp/cows/}"
 
 go run ./src/bin/pokedex/pokedex.go \
   -from "${FROM}" \
+  -fromMetadata "${FROM}pokemon.json" \
   -to ./build/assets/ \
   -toCategoryFpath pokedex.gob \
   -toDataSubDir cows/ \