--- a/test/pokesay_test.go
+++ b/test/pokesay_test.go
@@ -1,4 +1,4 @@
-package main
+package test
 
 import (
 	"testing"