… build/cows.tar.gz
… pokesay_test.go
… src/pokedex/entries.go
… src/pokesay/lookup.go
… src/pokesay/print.go
… src/pokesay/utils.go
… test/data/pokemon.json
… test/pokedex_test.go