## Context

This adds the japanese name of each pokemon as metadata, and an optional flag `-japanese-name` to trigger it.

```shell
 ☯ echo f | ./bin/pokesay-darwin-arm64 -name snorlax -japanese-name | tail -n 1

# The format is now
# > {name} | {japanese katakana} ({japanese pronunciation}) | {categories}
> snorlax | カビゴン (kabigon) | big/regular
```

In future I might amend this to be a two-liner if it proves to be too annoying on narrower terminal setups

## Implementation

There is a metadata file in the [pokesprite](https://github.com/msikma/pokesprite) repository: https://github.com/msikma/pokesprite/blob/c5aaa610ff2acdf7fd8e2dccd181bca8be9fcb3e/data/pokemon.json

<details>
  <summary>An example entry (Bulbasaur)</summary>

    {
      "idx": "001",
      "name": {
        "eng": "Bulbasaur",
        "chs": "妙蛙种子",
        "jpn": "フシギダネ",
        "jpn_ro": "Fushigidane"
      },
      "slug": {
        "eng": "bulbasaur",
        "jpn": "fushigidane",
        "jpn_ro": "fushigidane"
      },
      "gen-7": {
        "forms": {
          "$": {
            "has_female": false,
            "has_right": false
          }
        }
      },
      "gen-8": {
        "forms": {
          "$": {
            "is_prev_gen_icon": true
          }
        }
      }
    }

</details>

1. Download pokemon.json metadata file
  i. The Dockerfile now downloads this file
2. Build assets with new metadata
  i. the pokedex reads it in when creating the assets
  ii. adding the japanese names to the existing metadata files (1 for each individual cowfile)
3. Print the name if the flag `-japanese-name` is given

---

## Changes

- The Dockerfile addition, required jq to do some manipulation so that the data is easier to load via go
- Pokedex metadata method addition and refactor
  - split up metadata and category creation into 2 separate methods
- Add new flag to main pokesay file
  - refactor printing methods in pokesay, move to a subdir in src/

### Benchmarks

```shell
# New version with -japanese-name
 ☯ hyperfine 'echo f | ./bin/pokesay-linux-amd64 -japanese-name'
  Time (mean ± σ):       0.9 ms ±   0.1 ms    [User: 1.0 ms, System: 0.3 ms]
  Range (min … max):     0.6 ms …   1.6 ms    1181 runs

# Current version (0.9.0)
 ☯ hyperfine 'echo f | ~/bin/pokesay'
  Time (mean ± σ):       0.9 ms ±   0.1 ms    [User: 0.9 ms, System: 0.4 ms]
  Range (min … max):     0.5 ms …   1.5 ms    1126 runs
```

## TODOs

- [ ] Store name metadata separately
  - Storing all this metadata for each variation of a pokemon costs space (and possibly time?). 
- [ ] Improve print formatting
  - The parting joining by `|` is pretty janky, needs some love